+++ mobile/android/android-components/components/browser/engine-gecko-beta/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -266,6 +266,11 @@ class GeckoEngineSession(
         geckoSession.close()
     }
 
+    override fun recoverFromCrash(): Boolean {
+        // No-op: Functionality requires GeckoView 68.0
+        return false
+    }
+
     /**
      * NavigationDelegate implementation for forwarding callbacks to observers of the session.
      */