+++ mobile/android/android-components/components/browser/engine-servo/src/main/java/mozilla/components/browser/engine/servo/ServoEngineSession.kt
@@ -142,4 +142,9 @@ class ServoEngineSession(
     override fun exitFullScreenMode() {
         // not implemented yet
     }
+
+    override fun recoverFromCrash(): Boolean {
+        // not implemented yet
+        return false
+    }
 }